+++ CHANGELOG.md
@@ -14,6 +14,9 @@ changes, and bug fixes below. For complete details, see the git commit history.
 * Bug fixes and approx 50 % speed improvement, mainly from improvements to the hash map 
   part of the hash-heap data structures, improved register handling in context switches,
   and improved algorithms for some random number distributions.
+* Added functions to set the stack size of a `cmb_process`, either for that process 
+  only when initialized by `cmb_process_initialize_wssz()`, or globally for all future 
+  processes by `cmb_process_default_stacksize_set()`.
 * Added `cmb_priorityqueue`, `cmb_process_timer`, `cmb_process_yield()` and
   `cmb_process_resume()`.
 * Added `cimba_thread_hooks_set()` and `cimba_thread_context()` for managing CUDA streams.
@@ -23,7 +26,7 @@ changes, and bug fixes below. For complete details, see the git commit history.
   `cmb_logger_error()` (which abandons the current trial).
 * Adapted for use with ASan, UBSan, and TSan; and running these automatically on 
   GitHub CI runners after each git push.
-* Coroutines adapted for Windows 11 with its improved stack security measures.
+* Coroutines adapted for Windows 11 and modern CPUs with improved stack security measures.
 * Added `cmb_random_fmix64()` to bootstrap deterministic thread seeds from a master seed.
 * Added tutorial 5 demonstrating CUDA integration.